Refactor ContenFixerMigrator shortcodes functionality to ShortcodesMigrator
