+++ src/Command/WpCliCommands.php
@@ -24,6 +24,7 @@ class WpCliCommands {
 			PostsMigrator::class,
 			SettingsMigrator::class,
 			WooCommMigrator::class,
+			ShortcodesMigrator::class,
 		];
 
 		return apply_filters( 'newspack_migration_tools_command_classes', $classes_with_cli_commands );
